Added <TrackinInformation /> component
